+++ tardsquad_discord_bot/commands_rating.py
@@ -1,11 +1,10 @@
-from itertools import groupby
 from operator import attrgetter
 
 from discord.ext import commands
 
 import tardsquad_discord_bot
 from tardsquad_discord_bot.cog import TardBotCog
-from tardsquad_discord_bot.ratingentry import RatingEntry
+from tardsquad_discord_bot.ratingentry import RATING_EXPIRES_AFTER_H, RatingEntry
 from tardsquad_discord_bot.textcolor import TextColor
 
 
@@ -22,15 +21,16 @@ class TardBotRatingCommands(TardBotCog, name="Rating"):
         if not self.ratings:
             return "*Currently no active ratings collected*\n\nPlease share your rating with the `!rate` command."
         summary = []
-        sortkeyfn = attrgetter("rating")
-        for rating, entries in groupby(reversed(sorted(self.ratings.values(), key=sortkeyfn)), key=sortkeyfn):
-            names = [f"{entry.user.name} (expires in {entry.expires_at()})" for entry in entries]
-            summary.append(f"* {rating}: {', '.join(names)}")
+        for entry in reversed(sorted(self.ratings.values(), key=attrgetter("rating"))):
+            reason = f": {entry.reason}" if entry.reason else ""
+            summary.append(f"* {entry.rating} @{entry.user.name} [expires in {entry.expires_at()}]{reason}")
         summary_all = "\n".join(summary)
         return f"Current rating status:\n```{summary_all}```"
 
-    @commands.command(help="Rate how good you feel in a scale [1-10] 10 being best. Your rating is valid for 24h.")
-    async def rate(self, ctx, rating: int):
+    @commands.command(
+        help=f"Rate how good you feel on a scale [1-10], with 10 being best.\nYour rating is valid for {RATING_EXPIRES_AFTER_H}h.\nAn optional reason can also be given (must be a double-quoted string!)."
+    )
+    async def rate(self, ctx, rating: int, reason=None):
         reply = ""
         if rating < 1 or rating > 10:
             reply = TextColor.red("You can only be in states [1, 10].")
@@ -38,10 +38,10 @@ class TardBotRatingCommands(TardBotCog, name="Rating"):
             if rating == 10 and "fredrik" not in ctx.message.author.name.lower():
                 rating = 9
                 reply = TextColor.red("Sorry; a rating of 10 is reserved for Fredrik. I've downgraded you to a 9.")
-            rating_entry = RatingEntry(ctx.message.author, rating)
+            rating_entry = RatingEntry(ctx.message.author, rating, reason)
             self.ratings[ctx.message.author.id] = rating_entry
             # For testing with single user.
-            # self.ratings[str(ctx.message.author.id) + str(datetime.now())] = rating_entry
+            # self.ratings[str(ctx.message.author.id) + str(__import__("datetime").datetime.now())] = rating_entry
             reply += f":star: Thank you {ctx.message.author.name} for sharing your current rating!\n\n"
             reply += self._rating_status()
         await ctx.send(reply)