+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- An optional reason argument to the `!rate` command.
+- Handler for missing command argument. User will be encourage to check the help command.
+
 ### Changed
 - `!rate` expires after 6h instead of 24h.